+++ src/main/java/org/betonquest/betonquest/variables/PointVariable.java
@@ -1,10 +1,13 @@
 package org.betonquest.betonquest.variables;
 
+import lombok.CustomLog;
 import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.Instruction;
 import org.betonquest.betonquest.Point;
 import org.betonquest.betonquest.api.Variable;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
+import org.betonquest.betonquest.exceptions.ObjectNotFoundException;
+import org.betonquest.betonquest.id.ID;
 
 import java.util.List;
 import java.util.Locale;
@@ -14,27 +17,45 @@ import java.util.Locale;
  * some other amount.
  */
 @SuppressWarnings("PMD.CommentRequired")
+@CustomLog
 public class PointVariable extends Variable {
 
     protected String category;
     protected Type type;
     protected int amount;
 
+    @SuppressWarnings("PMD")
     public PointVariable(final Instruction instruction) throws InstructionParseException {
         super(instruction);
         category = instruction.next();
-        if (category.contains("*")) {
-            category = category.replace('*', '.');
-        } else {
-            category = instruction.getPackage().getName() + "." + category;
+
+        if (instruction.size() == 4) {
+            final String packagePath = instruction.current();
+            final String pointCategory = instruction.next();
+            try {
+                final ID id = new ID(instruction.getPackage(), packagePath + "." + pointCategory) {
+                };
+                category = id.getPackage().getName() + "." + pointCategory;
+            } catch (final ObjectNotFoundException e) {
+                LOG.warning(instruction.getPackage(), e.getMessage());
+            }
+
+        } else if (instruction.size() == 3) {
+            if (category.contains("*")) {
+                category = category.replace('*', '.');
+            } else {
+                category = instruction.getPackage().getName() + "." + category;
+            }
         }
+
+
         if ("amount".equalsIgnoreCase(instruction.next())) {
             type = Type.AMOUNT;
         } else if (instruction.current().toLowerCase(Locale.ROOT).startsWith("left:")) {
             type = Type.LEFT;
             try {
                 amount = Integer.parseInt(instruction.current().substring(5));
-            } catch (NumberFormatException e) {
+            } catch (final NumberFormatException e) {
                 throw new InstructionParseException("Could not parse point amount", e);
             }
         } else {